Took a vacation. As promised, the AST generator is now active. Its debug stream interface has been replaced with a lexer implementation for the job. There's more work to be done on it, but it can parse the passive one's test cases.
Some of the components of the builder return NULL, but no check is yet made by the caller for that return. This can and will lead to segfault until provisions are made, which they will be next commit. Some of the error handling is still in the wrong place, and must be moved such that the caller is doing the lion's share. Other than that, and a little tidying, and the AST module is ready for use.