Implements the class and struct keywords, as well as inheritence. Moves the read_type function out of context_parser: the relationship between the lexer, context, and coercer classes needs careful consideration. Cleans up some function definitions for Doxygen and adds lib to the README.
I need some method of unifying calls to read_type, et al, between the three main classes, while still allowing adopters to make calls to each system without having to instantiate a context. It might be frugal to use the global context as a sort of "default instance" of jdi::context for quick use of the evaluator class, while enabling sharing of intermediate functions. If this is what is to happen, redundant parameters in context_parser (namely, `lex') need removed.