git commit -m "Implements entering macro scalars. Makes macros copy-by-reference. Fixes invalid free in `llreader`. Improves error reporting info. Fixes "declaration with multiple types" error checking. Starts partitioning core lexer loops.
The parser is now capable of pushing the current open file onto a stack and replacing it with the value of a macro. I have also revamped that system in general.

I also changed the `copy(to)` method of `definition_scope` to `copy(from)` for consistency. I made `from` `const*`, and so added a typedef to the `const_iterator` members of both the member map and the macro map.

Macros themselves are now stored as `const macro_type*` to prevent modification of data. This is so macros can be copied by reference; all that is done is a ref count increment to prevent double free or early free.

The invalid free in `llreader` was caused by incorrect labeling of data acquired from an std::string in `llreader::encapsulate(string)`. The function labeled the open state as `FT_BUFFER`; the correct label was `FT_ALIAS`. These constants are now documented individually to prevent future confusion.

Tokens are now constructed with a legitimate filename (which there is presently no way to set from within the parser) as well as a line number and position relative to that line. The lattermost is done by tracking the position of the last line break and subtracting it from the current position on demand. I may keep track of file info when creating macros as well, to prevent mind injuries.

Error checking for declarations naming multiple primitives is now correct; `double int a` did not cause error last revision. This was due to primitives getting lumped in with modifier flags.

This commit takes a swing at getting rid of the redundant loops for skipping to the ends of strings and comments. This is good because it will not only save code, but allow quickly changing behavior of non-ISO C++ aspects such as escaped single-line comments.