Completes, to my knowledge, the AST builder, with exception of implementing a custom node for functions. Adds a precedence parameter to all parse_expression calls.
The parser can still not get through even basic C++ headers, but this is one less nail in the coffin. Before this commit, the parser got hung up on bool(_Sp::__value) in one of the type traits headers.

Evaluators for operators ::, ., ->, .*, and ->* must still be written. At very least, ::.