Starts moving all template handling code into handle_templates.
Some of it was in the class handling code, which is a big mistake.
Only enums should take any steps to account for templates, and even this should only go so far as optionally not evaluating on the spot.

In the next commit, enums will be fitted with a try-eval method wherein a failure to evaluate the expression does not constitute an error where an operand is a dependent value; figuring out exactly how to do this well will be a pain.