Problematic change: Add operator token types.
This change bloated up the lexer, a bit, but made it more readable. The lexer is about to undergo some major refactoring to fix a problem in the preprocessor, which seems to differ slightly from the standard.

This change likely breaks the main parser; I will perform more thorough diff testing after the tokens read from the GCC-preprocessed version of the engine are identical to the JDI-preprocessed tokens.