Start cleaning up, or at least spreading the mess around.
Delete some dead code.

Finish some half-assed or barely started code.
- Actually implement the push_buffer/pop_buffer routines.
- Clean up the way buffers were being treated.
- Band-aid the way the preprocessor's AST builder now interacts with type inference.
- Stitch up the macro handling code (but add more loose ends).

Next work item is to implement building the macro replacement list structural cache. Currently, the substitution logic requires it, but it's empty.

Another piece to pay attention to is the hellish mess that is read_type. That definitely needs some revisiting. Probably for C++11 dynamics, too, but also just to get a better grip on all the situations in which it's used. In particular, dependent types are handled poorly.