Overhaul jdi::value: use long double for all real types.
This isn't my favorite solution to a problem, but it's a portable way of dealing with unsigned longs without tons of extra handling.

On i386, generally 80-bit extended floats are supported, which have just enough significand for an unsigned long. On x64, floats are 128 bit and there's no problem at all.

Revisting this with a signed-magnitude representation for integers might be a better option. But for now, this enables checks like 0x7FFF'FFFF'FFFF'FFFFL < 0xFFFF'FFFF'FFFF'FFFFUL to work correctly in the preprocessor. I've added this as a unit test.