Fix whitespace handling in macros, mostly.
The build will now be green.

The current difference in behavior in JDI vs ISO is deliberate on my part. By standard, all whitespace in stringize operations is supposed to be replaced with a single space character. I include the entirety of the whitespace, as well as comments, in the hopes of improving debugging. I may need to correct this behavior later (it's a relatively simple fix).