Add GNU spellings of `__has_include`.
The error we were seeing was because GCC simply defined `__has_include` to be `__has_include__`, which isn't listed in the standard or the reference I was using for GNU extensions.

Same applies for `__include_next`.

Also, add a unit test for the string literal version of the expression, and modify the test for the non-string literal to attempt to botch the operation with a macro for the main filename component.