Fix GCC interactions.
The logic to automatically invoke GCC to print builtins and preprocess the ENIGMA engine was broken. This unifies a bunch of literal references and finishes automating the test end-to-end. Still can't be run on Travis because it depends on ENIGMA. Also, it'd be a brittle test; I'm sure more lexer features will become needed eventually.