Redo flag storage. This is not yet perfect.
In fact, this represents a small regression in the interests of a better path forward. Before, short and long were unique primitives; they weren't just short and long int. There was a bunch of logic to handle the implications of this in read_type, which really uglied up the code. Now, short and long are just modifiers on int, sharing a set of three flag bits.

After this change, `unsigned` and `signed` share a bit, so `signed` naturally means nothing. However, there is also a `mask` field, so `signed` and `unsigned` can be made to conflict with one another (and multiple uses of either can be rendered as an error).

Fixes the "two types named in declaration" problem we're seeing with `long long` (although these are currently not both recorded, as flag application is always an OR).

I haven't finished reasoning about how to handle `long long` in `read_type`. Thinking I'll just add a special case for it.