Fix lookup of C structs in scopes. Add unit test.
Checks in some new unit testing infrastructure to actually verify scope contents. Not advanced enough to verify the type flags from the other tests, yet, but way less brittle than asserting on a string representation of class that was parsed.

There may be more to this gift. This regression was added because I whimsically "corrected" JDI's storage of C structs so they wouldn't overlap with C++ definitions, per standard. We'll see if there are more regressions.